fix: find .npmrc by traversing parent directories
